backend/requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
django-cors-headers
qrcode
Pillow
